Move ScenarioProcessorTest.java and SnapshotRecordResultDeserialiserTest.java to the right package
